자재 인벤토리 컬럼 수정 및 도큐먼트파일 엔티티, DTO, 레퍼지토리 생성

+++ src/main/java/com/deligence/deli/repository/MaterialInventoryRepository.java
@@ -3,7 +3,7 @@ package com.deligence.deli.repository;
 import com.deligence.deli.domain.MaterialInventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialInventoryRepository extends JpaRepository<MaterialInventory, Long> {
+public interface MaterialInventoryRepository extends JpaRepository<MaterialInventory, Integer> {
